+++ R/datasets.R
@@ -364,6 +364,30 @@
 "movies_sample"
 
 
+#' International Power Lifting Results
+#' A subset of international powerlifting results.
+#' 
+#' @format A data frame with 41,152 entries, one entry for individual lifter
+#' \describe{
+#'   \item{name}{Individual lifter name}
+#'   \item{sex}{Binary sex (M/F)}
+#'   \item{event}{The type of competition that the lifter entered}
+#'   \item{equipment}{The equipment category under which the lifts were performed}
+#'   \item{age}{The age of the lifter on the start date of the meet}
+#'   \item{age_class}{The age class in which the filter falls}
+#'   \item{division}{division of competetition}
+#'   \item{bodyweight_kg}{The recorded bodyweight of the lifter at the time of competition, to two decimal places}
+#'   \item{weight_class_kg}{The weight class in which the lifter competed, to two decimal places}
+#'   \item{best3squat_kg}{Maximum of the first three successful attempts for the lift}
+#'   \item{best3bench_kg}{Maximum of the first three successful attempts for the lift}
+#'   \item{best3deadlift_kg}{Maximum of the first three successful attempts for the lift}
+#'   \item{place}{The recorded place of the lifter in the given division at the end of the meet}
+#'   \item{date}{Date of the event}
+#'   \item{federation}{The federation that hosted the meet}
+#'   \item{meet_name}{The name of the meet}
+#' }
+#' @source This data is a subset of the open dataset [Open Powerlifting](https://www.openpowerlifting.org/)
+"ipf_lifts"
 
 
 #' Data on maternal smoking and infant health
@@ -500,4 +524,4 @@
 #' \item{year}{Year of Sale}
 #' \item{region}{Region Where Sale Took Place}
 #' }
-"avocados" 
+"avocados"